pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "subtitle-forge"
version = "0.1.0"
description = "Local video subtitle generation and translation CLI tool"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["subtitle", "whisper", "translation", "cli", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "faster-whisper>=1.0.0",
    "ollama>=0.3.0",
    "ffmpeg-python>=0.2.0",
    "pysrt>=1.1.2",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
subtitle-forge = "subtitle_forge.cli.app:cli"

[project.urls]
Repository = "https://github.com/yourusername/subtitle-forge"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
